refactor: move the status feature out from under src/modules to src/
