+++ packages/api/src/app.module.ts
@@ -5,7 +5,7 @@ import { graphqlUploadExpress } from 'graphql-upload';
 import { LoggerMiddleware } from './common/logger.middleware';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { StatusModule } from './modules/status/status.module';
+import { StatusModule } from './status/status.module';
 import { UserModule } from './modules/user/user.module';
 import { AvatarModule } from './modules/avatar/avatar.module';
 import { AuthModule } from './modules/auth/auth.module';
